Update settings_navigation.md
While setting up my PKM where I document several plugins, it came to me that advanced URI could open the options page for any community plugin. It took a while to figure it out. The reason of my edit is to add a clearer example that would help the community to form the plugin URI.