fix: unexpected appending behavior when heading empty
Using `heading` and `append` with `separator` with a heading that is empty (not have any content in that heading block) could cause the content being inserted in the line where heading lies.

This commit fix the issue and will make sure a new line is always created not matter what separator user specified when appending to an empty heading.